+++ src/dataobject.h
@@ -17,7 +17,7 @@ class DataObject{
 public:
   /**** Create and Destory DataObject****/
   DataObject(DataType,
-    int blocksize=1, 
+    uint64_t blocksize=1, 
     int maxblock=4, 
     int memsize=4, 
     int outofcore=0,
